requirements.txt
# Postgres
psycopg2
# Django
Django==4.2.3
django-filter==23.2
django-environ==0.10.0
# Rest API
django-cors-headers==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-nested-routers==0.93.4
drf-spectacular==0.26.3
drf-spectacular-sidecar==2023.7.1

faker
